Refactor: Firebase integration and prompt management logic
- Moved Firebase admin initialization to a dedicated module for better organization.
- Updated import paths for Firebase admin and utility functions across various API routes.
- Replaced handleLazyMigration with a unified performLazyMigration function supporting batch and transaction modes.
- Consolidated prompt-related utility functions into promptUtils for improved maintainability.
- Introduced shared types for prompts to ensure consistency between front-end and back-end.
