+++ src/app/api/v1/prompts/[promptId]/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from 'next/server';
-import { adminDb } from '@/lib/firebaseAdmin';
+import { adminDb } from '@/server/db/firebase';
 
 export async function GET(
   req: Request,
