+++ src/server/utils/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Server 工具函式入口點
+ */
+
+export * from './promptUtils';
