+++ src/shared/index.ts
@@ -1,0 +1,5 @@
+/**
+ * 共用類型和常數的入口點
+ */
+
+export * from './types/prompt';
