+++ src/test/seqno-optimization-test.ts
@@ -7,7 +7,7 @@ import {
   normalizePromptSequence, 
   calculateInsertStrategy,
   type PromptData 
-} from '@/lib/utils/seqNoManager';
+} from '@/server/utils/promptUtils';
 
 export function runSeqNoTests() {
   console.log('🧪 開始 SeqNo 管理測試\n');
@@ -55,14 +55,14 @@ export function runSeqNoTests() {
   // 測試 2: calculateInsertStrategy - 插入到中間
   console.log('📋 測試 2: calculateInsertStrategy - 插入到 Prompt A 之後');
   try {
-    const { operations, insertSeqNo, affectedPrompts } = calculateInsertStrategy(
+    const { updateOperations, insertSeqNo, affectedPrompts } = calculateInsertStrategy(
       normalized, 
       '1' // Prompt A 的 ID
     );
     
     console.log('插入 seqNo:', insertSeqNo);
     console.log('受影響的 prompts:', affectedPrompts.map(p => p.name));
-    console.log('需要執行的操作:', operations.length);
+    console.log('需要執行的操作:', updateOperations.length);
     console.log('✅ 只有插入點之後的 prompt 會被影響\n');
   } catch (error) {
     console.error('❌ 測試失敗:', error);
@@ -71,13 +71,13 @@ export function runSeqNoTests() {
   // 測試 3: calculateInsertStrategy - 插入到最後
   console.log('📋 測試 3: calculateInsertStrategy - 插入到最後一個之後');
   try {
-    const { operations, insertSeqNo } = calculateInsertStrategy(
+    const { updateOperations, insertSeqNo } = calculateInsertStrategy(
       normalized, 
       normalized[normalized.length - 1].id
     );
     
     console.log('插入 seqNo:', insertSeqNo);
-    console.log('受影響的操作數:', operations.length);
+    console.log('受影響的操作數:', updateOperations.length);
     console.log('✅ 插入到最後時，不應該有其他 prompt 受影響\n');
   } catch (error) {
     console.error('❌ 測試失敗:', error);
@@ -115,13 +115,13 @@ export function runPerformanceComparison() {
 
   // 測試插入策略效能
   console.time('插入策略計算');
-  const { operations } = calculateInsertStrategy(largePromptSet, 'prompt-50');
+  const { updateOperations } = calculateInsertStrategy(largePromptSet, 'prompt-50');
   console.timeEnd('插入策略計算');
   
   console.log('📊 結果統計:');
   console.log(`- 總 prompts 數: ${largePromptSet.length}`);
-  console.log(`- 受影響的 prompts: ${operations.length}`);
-  console.log(`- 效能提升: ${Math.round((1 - operations.length / largePromptSet.length) * 100)}% 的 prompts 不需要更新`);
+  console.log(`- 受影響的 prompts: ${updateOperations.length}`);
+  console.log(`- 效能提升: ${Math.round((1 - updateOperations.length / largePromptSet.length) * 100)}% 的 prompts 不需要更新`);
   
   console.log('\n✅ 新方案只更新必要的 prompts，大幅提升效能');
 }
